Renamed allreduce file
