--- ompi/mca/coll/libpnbc_osc/pnbc_osc_rallreduce.c
@@ -1,910 +1,0 @@
-/* -*- Mode: C; c-basic-offset:2 ; indent-tabs-mode:nil -*- */
-/*
- * Copyright (c) 2006      The Trustees of Indiana University and Indiana
- *                         University Research and Technology
- *                         Corporation.  All rights reserved.
- * Copyright (c) 2006      The Technical University of Chemnitz. All
- *                         rights reserved.
- * Copyright (c) 2013-2017 Los Alamos National Security, LLC. All rights
- *                         reserved.
- * Copyright (c) 2014-2018 Research Organization for Information Science
- *                         and Technology (RIST).  All rights reserved.
- * Copyright (c) 2017      IBM Corporation.  All rights reserved.
- * Copyright (c) 2018      FUJITSU LIMITED.  All rights reserved.
- * $COPYRIGHT$
- *
- * Additional copyrights may follow
- *
- * Author(s): Torsten Hoefler <htor@cs.indiana.edu>
- *
- */
-#include "pnbc_osc_internal.h"
-#include "ompi/communicator/communicator.h"
-#include "ompi/datatype/ompi_datatype.h"
-#include "ompi/op/op.h"
-#include "opal/util/bit_ops.h"
-
-#include <assert.h>
-
-static inline int allred_sched_diss(int rank, int p, int count, MPI_Datatype datatype, ptrdiff_t gap,
-                                    void *sendbuf, void *recvbuf, MPI_Op op, char inplace,
-                                    NBC_Schedule *schedule, void *tmpbuf, ompi_request_t ** request);
-static inline int allred_sched_ring(int rank, int p, int count, MPI_Datatype datatype,
-                                    const void *sendbuf, void *recvbuf, MPI_Op op, int size, int ext,
-                                    NBC_Schedule *schedule, void *tmpbuf);
-static inline int allred_sched_linear(int rank, int p, const void *sendbuf, void *recvbuf, int count,
-                                      MPI_Datatype datatype, ptrdiff_t gap, MPI_Op op, int ext,
-                                      int size, NBC_Schedule *schedule, void *tmpbuf);
-
-static inline int allred_sched_redscat_allgather( int rank, int comm_size, int count,
-                                                  MPI_Datatype datatype, ptrdiff_t gap,
-                                                  const void *sbuf, void *rbuf, MPI_Op op,
-                                                  char inplace, NBC_Schedule *schedule,
-                                                  void *tmpbuf, struct ompi_communicator_t *comm);
-
-
-static int nbc_rallreduce_init(void* sendbuf, void* recvbuf, int count, MPI_Datatype datatype,
-                               MPI_Op op, struct ompi_communicator_t *comm, ompi_request_t **request,
-                               struct mca_coll_base_module_2_3_0_t *module, bool persistent)
-{
-  int rank, p, res;
-  ptrdiff_t ext, lb;
-  NBC_Schedule *schedule;
-  size_t size;
-  ompi_win_t *win;
-  opal_info_t *info = NULL;
-  enum { NBC_ARED_BINOMIAL, NBC_ARED_RING, NBC_ARED_REDSCAT_ALLGATHER } alg;
-  char inplace;
-  void *tmpbuf = NULL;
-  ompi_coll_libpnbc_osc_module_t *libpnbc_osc_module = (ompi_coll_libpnbc_osc_module_t*) module;
-  ptrdiff_t span, gap;
-  MPI_Aint disp, *disp_a;
-
-  NBC_IN_PLACE(sendbuf, recvbuf, inplace);
-
-  rank = ompi_comm_rank (comm);
-  p = ompi_comm_size (comm);
-
-  /* checks for datatype extent */
-  res = ompi_datatype_get_extent(datatype, &lb, &ext);
-  if (OMPI_SUCCESS != res) {
-    NBC_Error ("MPI Error in ompi_datatype_type_extent() (%i)", res);
-    return res;
-  }
-  /* checks for datatype size */
-  res = ompi_datatype_type_size (datatype, &size);
-  if (OMPI_SUCCESS != res) {
-    NBC_Error ("MPI Error in ompi_datatype_type_size() (%i)", res);
-    return res;
-  }
-
-  if (1 == p && (!persistent || inplace)) {
-    if (!inplace) {
-      /* for a single node - copy data to receivebuf */
-      res = NBC_Copy(sendbuf, count, datatype, recvbuf, count, datatype, comm);
-      if (OPAL_UNLIKELY(OMPI_SUCCESS != res)) {
-        return res;
-      }
-    }
-    return nbc_get_noop_request(persistent, request);
-  }
-
-  span = opal_datatype_span(&datatype->super, count, &gap);
-  tmpbuf = malloc (span);
-  if (OPAL_UNLIKELY(NULL == tmpbuf)) {
-    return OMPI_ERR_OUT_OF_RESOURCE;
-  }
-
-  /* create an MPI dynamic Window */
-  res = ompi_win_create_dynamic(info, comm, &win);
-  if (OMPI_SUCCESS != res) {
-    NBC_Error ("MPI Error in win_create_dynamic (%i)", res);
-    return res;
-  }
-
-  /* Attach window to sendbuf */
-  res = win->w_osc_module->osc_win_attach(win, sendbuf, count*size);
-
-  /* MPI Get_address  */
-  MPI_Get_address (sendbuf, &disp);
-
-  /* create an array of displacements */
-  disp_a = (MPI_Aint*)malloc(count * sizeof(MPI_Aint));
-
-  /* Gather ALL displacements-> pt2pt call */
-
-  res = comm->c_coll->coll_allgather(&disp, 1, MPI_INT, disp_a, count, MPI_INT, comm,
-                                     comm->c_coll->coll_allgather_module);
-  if (OMPI_SUCCESS != res) {
-    NBC_Error ("MPI Error in coll_allgather (%i)", res);
-    return res;
-  }
-
-  /* algorithm selection - NOTE: implemented binomial alg for now only */
-  int nprocs_pof2 = opal_next_poweroftwo(p) >> 1;
-  if (libpnbc_osc_iallreduce_algorithm == 0) {
-    if(p < 4 || size*count < 65536 || !ompi_op_is_commute(op) || inplace) {
-      alg = NBC_ARED_BINOMIAL;
-    } else if (count >= nprocs_pof2 && ompi_op_is_commute(op)) {
-      alg = NBC_ARED_REDSCAT_ALLGATHER;
-    } else {
-      alg = NBC_ARED_RING;
-    }
-  } else {
-    if (libpnbc_osc_iallreduce_algorithm == 1)
-      alg = NBC_ARED_RING;
-    else if (libpnbc_osc_iallreduce_algorithm == 2)
-      alg = NBC_ARED_BINOMIAL;
-    else if (libpnbc_osc_iallreduce_algorithm == 3 && count >= nprocs_pof2 && ompi_op_is_commute(op)) {
-      alg = NBC_ARED_REDSCAT_ALLGATHER;
-    } else
-      alg = NBC_ARED_RING;
-  }
-
-  /* create a new schedule */
-  schedule = OBJ_NEW(NBC_Schedule);
-  if (NULL == schedule) {
-    free(tmpbuf);
-    return OMPI_ERR_OUT_OF_RESOURCE;
-  }
-
-  if (p == 1) {
-    res = NBC_Sched_copy((void *)sendbuf, false, count, datatype,
-                         recvbuf, false, count, datatype, schedule, false);
-  } else {
-    switch(alg) {
-      case NBC_ARED_BINOMIAL:
-        res = allred_sched_diss(rank, p, count, datatype, gap, sendbuf, recvbuf, op, inplace,
-                                schedule, tmpbuf, request);
-        break;
-      case NBC_ARED_REDSCAT_ALLGATHER:
-        res = allred_sched_redscat_allgather(rank, p, count, datatype, gap, sendbuf, recvbuf,
-                                             op, inplace, schedule, tmpbuf, comm);
-        break;
-      case NBC_ARED_RING:
-        res = allred_sched_ring(rank, p, count, datatype, sendbuf, recvbuf, op, size, ext, schedule,
-                                tmpbuf);
-        break;
-    }
-  }
-
-  res = NBC_Sched_win_free(schedule, true);
-  if (OPAL_UNLIKELY(OMPI_SUCCESS != res)) {
-    OBJ_RELEASE(schedule);
-    free(tmpbuf);
-    return res;
-  }
-
-  if (OPAL_UNLIKELY(OMPI_SUCCESS != res)) {
-    OBJ_RELEASE(schedule);
-    free(tmpbuf);
-    return res;
-  }
-
-  res = NBC_Sched_commit(schedule);
-  if (OPAL_UNLIKELY(OMPI_SUCCESS != res)) {
-    OBJ_RELEASE(schedule);
-    free(tmpbuf);
-    return res;
-  }
-
-
-  res = NBC_Schedule_request (schedule, comm, libpnbc_osc_module, persistent, request, tmpbuf);
-  if (OPAL_UNLIKELY(OMPI_SUCCESS != res)) {
-    OBJ_RELEASE(schedule);
-    free(tmpbuf);
-    return res;
-  }
-
-  return OMPI_SUCCESS;
-}
-
-int ompi_coll_libpnbc_osc_rallreduce(void* sendbuf, void* recvbuf, int count, MPI_Datatype datatype,
-                                     MPI_Op op, struct ompi_communicator_t *comm,
-                                     ompi_request_t ** request,
-                                     struct mca_coll_base_module_2_3_0_t *module) {
-  
-  int res = nbc_rallreduce_init(sendbuf, recvbuf, count, datatype, op, comm, request, module, false);
-  if (OPAL_UNLIKELY(OMPI_SUCCESS != res)) {
-    return res;
-  }
-
-  res = NBC_Start(*(ompi_coll_libpnbc_osc_request_t **)request);
-  if (OPAL_UNLIKELY(OMPI_SUCCESS != res)) {
-    NBC_Return_handle (*(ompi_coll_libpnbc_osc_request_t **)request);
-    *request = &ompi_request_null.request;
-    return res;
-  }
-
-  return OMPI_SUCCESS;
-}
-
-/* binomial allreduce (binomial tree up and binomial bcast down)
- * working principle:
- * - each node gets a virtual rank vrank
- * - the 'root' node get vrank 0
- * - node 0 gets the vrank of the 'root'
- * - all other ranks stay identical (they do not matter)
- *
- * Algorithm:
- * pairwise exchange
- * round r:
- *  grp = rank % 2^r
- *  if grp == 0: receive from rank + 2^(r-1) if it exists and reduce value
- *  if grp == 1: send to rank - 2^(r-1) and exit function
- *
- * do this for R=log_2(p) rounds
- * followed by a Bcast:
- * Algorithm:
- * - each node with vrank > 2^r and vrank < 2^r+1 receives from node
- *   vrank - 2^r (vrank=1 receives from 0, vrank 0 receives never)
- * - each node sends each round r to node vrank + 2^r
- * - a node stops to send if 2^r > commsize
- *
- */
-#define RANK2VRANK(rank, vrank, root)           \
-  {                                             \
-    vrank = rank;                               \
-    if (rank == 0) vrank = root;                \
-    if (rank == root) vrank = 0;                \
-  }
-#define VRANK2RANK(rank, vrank, root)           \
-  {                                             \
-    rank = vrank;                               \
-    if (vrank == 0) rank = root;                \
-    if (vrank == root) rank = 0;                \
-  }
-
-static inline int allred_sched_diss(int rank, int p, int count, MPI_Datatype datatype, ptrdiff_t gap,
-                                    void *sendbuf, void *recvbuf, MPI_Op op, char inplace,
-                                    NBC_Schedule *schedule, void *tmpbuf, ompi_request_t **request) {
-
-  int root, vrank, maxr, vpeer, peer, res;
-
-
-  int lock_type = MPI_LOCK_SHARED;
-  int assert = 0;
-  root = 0; /* this makes the code for ireduce and iallreduce nearly identical
-               - could be changed to improve performance */
-
-  RANK2VRANK(rank, vrank, root);
-  maxr = (int)ceil((log((double)p)/LOG2));
-
-  /* form a binary tree, try_get and add from children */
-  for (int r = 1; r <= maxr ; ++r) {
-    if ((vrank % (1 << r)) == 0) {
-      /* we have to receive this round */
-      vpeer = vrank + (1 << (r - 1));
-      VRANK2RANK(peer, vpeer, root);
-      if (peer < p) {
-        /* get the data from my peer and store it in recvbuf*/
-        res = NBC_Sched_try_get (recvbuf, false, count, datatype, peer, count, datatype,
-                                 schedule, lock_type, assert, true);
-        if (OPAL_UNLIKELY(OMPI_SUCCESS != res)) {
-          return res;
-        }
-        res = NBC_Sched_barrier (schedule);
-        if (OPAL_UNLIKELY(OMPI_SUCCESS != res)) {
-          return res;
-        }
-        /* add value to my values in my window - i.e. add the values and store them in sendbuf
-           so they can be use in the next r iteration */
-        res = NBC_Sched_op (recvbuf, false, sendbuf, false, count, datatype, op, schedule, true);
-        if (OPAL_UNLIKELY(OMPI_SUCCESS != res)) {
-          return res;
-        }
-      }
-    }
-  }
-
-  /* this is the Bcast part */
-  RANK2VRANK(rank, vrank, root);
-  
-  /* receive from the right hosts  */
-  if (vrank != 0) {
-    for (int r = 0; r < maxr ; ++r) {
-      if ((vrank >= (1 << r)) && (vrank < (1 << (r + 1)))) {
-        VRANK2RANK(peer, vrank - (1 << r), root);
-        /* try_get and place in recvbuf*/
-        res = NBC_Sched_try_get (recvbuf, false, count, datatype, peer, count, datatype,
-                                 schedule, lock_type, assert, true);
-        if (OPAL_UNLIKELY(OMPI_SUCCESS != res)) {
-          return res;
-        }
-        
-      }
-    }
-  }
-
-  /* root copies buffer to recvbuf */
-  if (0 == vrank){
-    res = NBC_Sched_copy(sendbuf, false, count, datatype, recvbuf, false, count, datatype,
-                         schedule, false);
-    if (OPAL_UNLIKELY(OMPI_SUCCESS != res)) {
-      return res;
-    }
-
-  }
-    /* end of the bcast */
-    return OMPI_SUCCESS;
-  }
-
-  static inline int allred_sched_ring (int r, int p, int count, MPI_Datatype datatype,
-                                       const void *sendbuf, void *recvbuf, MPI_Op op,
-                                       int size, int ext, NBC_Schedule *schedule, void *tmpbuf) {
-    int segsize, *segsizes, *segoffsets; /* segment sizes and offsets per segment (number of segments == number of nodes */
-    int speer, rpeer; /* send and recvpeer */
-    int res = OMPI_SUCCESS;
-
-    if (count == 0) {
-      return OMPI_SUCCESS;
-    }
-
-    segsizes = (int *) malloc (sizeof (int) * p);
-    segoffsets = (int *) malloc (sizeof (int) * p);
-    if (NULL == segsizes || NULL == segoffsets) {
-      free (segsizes);
-      free (segoffsets);
-      return OMPI_ERR_OUT_OF_RESOURCE;
-    }
-
-    segsize = (count + p - 1) / p; /* size of the segments */
-
-    segoffsets[0] = 0;
-    for (int i = 0, mycount = count ; i < p ; ++i) {
-      mycount -= segsize;
-      segsizes[i] = segsize;
-      if (mycount < 0) {
-        segsizes[i] = segsize + mycount;
-        mycount = 0;
-      }
-
-      if (i) {
-        segoffsets[i] = segoffsets[i-1] + segsizes[i-1];
-      }
-    }
-
-    /* reduce peers */
-    speer = (r + 1) % p;
-    rpeer = (r - 1 + p) % p;
-
-    /*  + -> reduced this round
-     *  / -> sum (reduced in a previous step)
-     *
-     *     *** round 0 ***
-     *    0        1        2
-     *
-     *   00       10       20      0: [1] -> 1
-     *   01       11       21      1: [2] -> 2
-     *   02       12       22      2: [0] -> 0  --> send element (r+1)%p to node (r+1)%p
-     *
-     *      *** round 1 ***
-     *    0        1        2
-     *
-     *   00+20    10       20     0: red(0), [0] -> 1
-     *   01       11+01    21     1: red(1), [1] -> 2
-     *   02       12       22+12  2: red(2), [2] -> 0 --> reduce and send element (r+0)%p to node (r+1)%p
-     *
-     *      *** round 2 ***
-     *    0        1        2
-     *
-     *   00/20    all      20     0: red(2), [2] -> 1
-     *   01       11/01    all    1: red(0), [0] -> 2
-     *   all      12       22/12  2: red(1), [1] -> 0 --> reduce and send (r-1)%p to node (r+1)%p
-     *
-     *      *** round 3 ***
-     *    0        1        2
-     *
-     *   00/20    all      all    0: [1] -> 1
-     *   all      11/01    all    1: [2] -> 2
-     *   all      all      22/12  2: [0] -> 0 --> send element (r-2)%p to node (r+1)%p
-     *
-     *      *** round 4 ***
-     *    0        1        2
-     *
-     *   all      all      all    0: done
-     *   all      all      all    1: done
-     *   all      all      all    2: done
-     *
-     * -> 4
-     *     *** round 0 ***
-     *    0        1        2        3
-     *
-     *   00       10       20       30       0: [1] -> 1
-     *   01       11       21       31       1: [2] -> 2
-     *   02       12       22       32       2: [3] -> 3
-     *   03       13       23       33       3: [0] -> 0 --> send element (r+1)%p to node (r+1)%p
-     *
-     *      *** round 1 ***
-     *    0        1        2        3
-     *
-     *   00+30    10       20       30       0: red(0), [0] -> 1
-     *   01       11+01    21       31       1: red(1), [1] -> 2
-     *   02       12       22+12    32       2: red(2), [2] -> 3
-     *   03       13       23       33+23    3: red(3), [3] -> 0 --> reduce and send element (r+0)%p to node (r+1)%p
-     *
-     *      *** round 2 ***
-     *    0        1        2        3
-     *
-     *   00/30    10+00/30 20       30       0: red(3), [3] -> 1
-     *   01       11/01    21+11/01 31       1: red(0), [0] -> 2
-     *   02       12       22/12    32+22/12 2: red(1), [1] -> 3
-     *   03+33/23 13       23       33/23    3: red(2), [2] -> 0 --> reduce and send (r-1)%p to node (r+1)%p
-     *
-     *      *** round 3 ***
-     *    0        1        2        3
-     *
-     *   00/30    10/00/30 all      30       0: red(2), [2] -> 1
-     *   01       11/01    21/11/01 all      1: red(3), [3] -> 2
-     *   all      12       22/12    32/22/12 2: red(0), [0] -> 3
-     *   03/33/23 all      23       33/23    3: red(1), [1] -> 0 --> reduce and send (r-2)%p to node (r+1)%p
-     *
-     *      *** round 4 ***
-     *    0        1        2        3
-     *
-     *   00/30    10/00/30 all      all      0: [1] -> 1
-     *   all      11/01    21/11/01 all      1: [2] -> 2
-     *   all      all      22/12    32/22/12 2: [3] -> 3
-     *   03/33/23 all      all      33/23    3: [0] -> 0 --> receive and send element (r+1)%p to node (r+1)%p
-     *
-     *      *** round 5 ***
-     *    0        1        2        3
-     *
-     *   all      10/00/30 all      all      0: [0] -> 1
-     *   all      all      21/11/01 all      1: [1] -> 2
-     *   all      all      all      32/22/12 2: [3] -> 3
-     *   03/33/23 all      all      all      3: [4] -> 4 --> receive and send element (r-0)%p to node (r+1)%p
-     *
-     *      *** round 6 ***
-     *    0        1        2        3
-     *
-     *   all      all      all      all
-     *   all      all      all      all
-     *   all      all      all      all
-     *   all      all      all      all     receive element (r-1)%p
-     *
-     *   2p-2 rounds ... every node does p-1 reductions and p-1 sends
-     *
-     */
-    /* first p-1 rounds are reductions */
-    for (int round = 0 ; round < p - 1 ; ++round) {
-      int selement = (r+1-round + 2*p /*2*p avoids negative mod*/)%p; /* the element I am sending */
-      int soffset = segoffsets[selement]*ext;
-      int relement = (r-round + 2*p /*2*p avoids negative mod*/)%p; /* the element that I receive from my neighbor */
-      int roffset = segoffsets[relement]*ext;
-
-      /* first message come out of sendbuf */
-      if (round == 0) {
-        res = NBC_Sched_send ((char *) sendbuf + soffset, false, segsizes[selement], datatype, speer,
-                              schedule, false);
-      } else {
-        res = NBC_Sched_send ((char *) recvbuf + soffset, false, segsizes[selement], datatype, speer,
-                              schedule, false);
-      }
-
-      if (OPAL_UNLIKELY(OMPI_SUCCESS != res)) {
-        break;
-      }
-
-      res = NBC_Sched_recv ((char *) recvbuf + roffset, false, segsizes[relement], datatype, rpeer,
-                            schedule, true);
-      if (OPAL_UNLIKELY(OMPI_SUCCESS != res)) {
-        break;
-      }
-
-      res = NBC_Sched_op ((char *) sendbuf + roffset, false, (char *) recvbuf + roffset, false,
-                          segsizes[relement], datatype, op, schedule, true);
-      if (OPAL_UNLIKELY(OMPI_SUCCESS != res)) {
-        break;
-      }
-    }
-
-    if (OPAL_UNLIKELY(OMPI_SUCCESS != res)) {
-      free (segsizes);
-      free (segoffsets);
-      return res;
-    }
-
-    for (int round = p - 1 ; round < 2 * p - 2 ; ++round) {
-      int selement = (r+1-round + 2*p /*2*p avoids negative mod*/)%p; /* the element I am sending */
-      int soffset = segoffsets[selement]*ext;
-      int relement = (r-round + 2*p /*2*p avoids negative mod*/)%p; /* the element that I receive from my neighbor */
-      int roffset = segoffsets[relement]*ext;
-
-      res = NBC_Sched_send ((char *) recvbuf + soffset, false, segsizes[selement], datatype, speer,
-                            schedule, false);
-      if (OPAL_UNLIKELY(OMPI_SUCCESS != res)) {
-        break;
-      }
-
-      res = NBC_Sched_recv ((char *) recvbuf + roffset, false, segsizes[relement], datatype, rpeer,
-                            schedule, true);
-      if (OPAL_UNLIKELY(OMPI_SUCCESS != res)) {
-        break;
-      }
-    }
-
-    free (segsizes);
-    free (segoffsets);
-
-    return res;
-  }
-
-  static inline int allred_sched_linear(int rank, int rsize, const void *sendbuf, void *recvbuf,
-                                        int count, MPI_Datatype datatype, ptrdiff_t gap, MPI_Op op,
-                                        int ext, int size, NBC_Schedule *schedule, void *tmpbuf) {
-    int res;
-
-    if (0 == count) {
-      return OMPI_SUCCESS;
-    }
-
-    /* send my data to the remote root */
-    res = NBC_Sched_send (sendbuf, false, count, datatype, 0, schedule, false);
-    if (OPAL_UNLIKELY(OMPI_SUCCESS != res)) {
-      return res;
-    }
-
-    /* recv my data to the remote root */
-    if (0 != rank || 1 ==(rsize%2)) {
-      res = NBC_Sched_recv (recvbuf, false, count, datatype, 0, schedule, false);
-    } else {
-      res = NBC_Sched_recv ((void *)(-gap), true, count, datatype, 0, schedule, false);
-    }
-    if (OPAL_UNLIKELY(OMPI_SUCCESS != res)) {
-      return res;
-    }
-
-    if (0 == rank) {
-      char *rbuf, *lbuf, *buf;
-      int tmprbuf, tmplbuf;
-
-      res = NBC_Sched_barrier (schedule);
-      if (OPAL_UNLIKELY(OMPI_SUCCESS != res)) {
-        return res;
-      }
-
-      /* ensure the result ends up in recvbuf */
-      if (0 == (rsize%2)) {
-        lbuf = (void *)(-gap);
-        tmplbuf = true;
-        rbuf = recvbuf;
-        tmprbuf = false;
-      } else {
-        rbuf = (void *)(-gap);
-        tmprbuf = true;
-        lbuf = recvbuf;
-        tmplbuf = false;
-      }
-
-      /* get data from remote peers and reduce */
-      for (int rpeer = 1 ; rpeer < rsize ; ++rpeer) {
-        res = NBC_Sched_recv (rbuf, tmprbuf, count, datatype, rpeer, schedule, true);
-        if (OPAL_UNLIKELY(OMPI_SUCCESS != res)) {
-          return res;
-        }
-
-        res = NBC_Sched_op (lbuf, tmplbuf, rbuf, tmprbuf, count, datatype, op, schedule, true);
-        if (OPAL_UNLIKELY(OMPI_SUCCESS != res)) {
-          return res;
-        }
-        /* swap left and right buffers */
-        buf = rbuf; rbuf = lbuf ; lbuf = buf;
-        tmprbuf ^= 1; tmplbuf ^= 1;
-      }
-
-      /* exchange our result with the remote root (each root will broadcast to the other's peers) */
-      res = NBC_Sched_recv ((void *)(-gap), true, count, datatype, 0, schedule, false);
-      if (OPAL_UNLIKELY(OMPI_SUCCESS != res)) {
-        return res;
-      }
-
-      /* wait for data from remote root */
-      res = NBC_Sched_send (recvbuf, false, count, datatype, 0, schedule, true);
-      if (OPAL_UNLIKELY(OMPI_SUCCESS != res)) {
-        return res;
-      }
-
-      /* broadcast the result to all remote peers */
-      for (int rpeer = 1 ; rpeer < rsize ; ++rpeer) {
-        res = NBC_Sched_send ((void *)(-gap), true, count, datatype, rpeer, schedule, false);
-        if (OPAL_UNLIKELY(OMPI_SUCCESS != res)) {
-          return res;
-        }
-      }
-    }
-
-    return OMPI_SUCCESS;
-  }
-
-  /*
-   * allred_sched_redscat_allgather:
-   *
-   * Description: an implementation of Rabenseifner's Allreduce algorithm [1, 2].
-   *   [1] Rajeev Thakur, Rolf Rabenseifner and William Gropp.
-   *       Optimization of Collective Communication Operations in MPICH //
-   *       The Int. Journal of High Performance Computing Applications. Vol 19,
-   *       Issue 1, pp. 49--66.
-   *   [2] http://www.hlrs.de/mpi/myreduce.html.
-   *
-   * This algorithm is a combination of a reduce-scatter implemented with
-   * recursive vector halving and recursive distance doubling, followed either
-   * by an allgather implemented with recursive doubling.
-   *
-   * Step 1. If the number of processes is not a power of two, reduce it to
-   * the nearest lower power of two (p' = 2^{\floor{\log_2 p}})
-   * by removing r = p - p' extra processes as follows. In the first 2r processes
-   * (ranks 0 to 2r - 1), all the even ranks send the second half of the input
-   * vector to their right neighbor (rank + 1), and all the odd ranks send
-   * the first half of the input vector to their left neighbor (rank - 1).
-   * The even ranks compute the reduction on the first half of the vector and
-   * the odd ranks compute the reduction on the second half. The odd ranks then
-   * send the result to their left neighbors (the even ranks). As a result,
-   * the even ranks among the first 2r processes now contain the reduction with
-   * the input vector on their right neighbors (the odd ranks). These odd ranks
-   * do not participate in the rest of the algorithm, which leaves behind
-   * a power-of-two number of processes. The first r even-ranked processes and
-   * the last p - 2r processes are now renumbered from 0 to p' - 1.
-   *
-   * Step 2. The remaining processes now perform a reduce-scatter by using
-   * recursive vector halving and recursive distance doubling. The even-ranked
-   * processes send the second half of their buffer to rank + 1 and the odd-ranked
-   * processes send the first half of their buffer to rank - 1. All processes
-   * then compute the reduction between the local buffer and the received buffer.
-   * In the next log_2(p') - 1 steps, the buffers are recursively halved, and the
-   * distance is doubled. At the end, each of the p' processes has 1 / p' of the
-   * total reduction result.
-   *
-   * Step 3. An allgather is performed by using recursive vector doubling and
-   * distance halving. All exchanges are executed in reverse order relative
-   * to recursive doubling on previous step. If the number of processes is not
-   * a power of two, the total result vector must be sent to the r processes
-   * that were removed in the first step.
-   *
-   * Limitations:
-   *   count >= 2^{\floor{\log_2 p}}
-   *   commutative operations only
-   *   intra-communicators only
-   *
-   * Memory requirements (per process):
-   *   count * typesize + 4 * \log_2(p) * sizeof(int) = O(count)
-   *
-   * Schedule length (rounds): O(\log(p))
-   */
-  static inline int allred_sched_redscat_allgather(int rank, int comm_size, int count,
-                                                   MPI_Datatype datatype, ptrdiff_t gap,
-                                                   const void *sbuf, void *rbuf, MPI_Op op, char inplace,
-                                                   NBC_Schedule *schedule, void *tmpbuf,
-                                                   struct ompi_communicator_t *comm)
-  {
-    int res = OMPI_SUCCESS;
-    int *rindex = NULL, *rcount = NULL, *sindex = NULL, *scount = NULL;
-    /* Find nearest power-of-two less than or equal to comm_size */
-    int nsteps = opal_hibit(comm_size, comm->c_cube_dim + 1);   /* ilog2(comm_size) */
-    int nprocs_pof2 = 1 << nsteps;                              /* flp2(comm_size) */
-    if (!inplace) {
-      res = NBC_Sched_copy((char *)sbuf, false, count, datatype,
-                           rbuf, false, count, datatype, schedule, true);
-      if (OPAL_UNLIKELY(OMPI_SUCCESS != res)) { goto cleanup_and_return; }
-    }
-    char *tmp_buf = (char *)tmpbuf - gap;
-    ptrdiff_t lb, extent;
-    ompi_datatype_get_extent(datatype, &lb, &extent);
-    /*
-     * Step 1. Reduce the number of processes to the nearest lower power of two
-     * p' = 2^{\floor{\log_2 p}} by removing r = p - p' processes.
-     * 1. In the first 2r processes (ranks 0 to 2r - 1), all the even ranks send
-     *    the second half of the input vector to their right neighbor (rank + 1)
-     *    and all the odd ranks send the first half of the input vector to their
-     *    left neighbor (rank - 1).
-     * 2. All 2r processes compute the reduction on their half.
-     * 3. The odd ranks then send the result to their left neighbors
-     *    (the even ranks).
-     *
-     * The even ranks (0 to 2r - 1) now contain the reduction with the input
-     * vector on their right neighbors (the odd ranks). The first r even
-     * processes and the p - 2r last processes are renumbered from
-     * 0 to 2^{\floor{\log_2 p}} - 1.
-     */
-    int vrank, step, wsize;
-    int nprocs_rem = comm_size - nprocs_pof2;
-    if (rank < 2 * nprocs_rem) {
-      int count_lhalf = count / 2;
-      int count_rhalf = count - count_lhalf;
-      if (rank % 2 != 0) {
-        /*
-         * Odd process -- exchange with rank - 1
-         * Send the left half of the input vector to the left neighbor,
-         * Recv the right half of the input vector from the left neighbor
-         */
-        res = NBC_Sched_send(rbuf, false, count_lhalf, datatype, rank - 1,
-                             schedule, false);
-        if (OPAL_UNLIKELY(OMPI_SUCCESS != res)) { goto cleanup_and_return; }
-        res = NBC_Sched_recv(tmp_buf + (ptrdiff_t)count_lhalf * extent,
-                             false, count_rhalf, datatype, rank - 1, schedule, true);
-        if (OPAL_UNLIKELY(OMPI_SUCCESS != res)) { goto cleanup_and_return; }
-        res = NBC_Sched_op(tmp_buf + (ptrdiff_t)count_lhalf * extent,
-                           false, (char *)rbuf + (ptrdiff_t)count_lhalf * extent,
-                           false, count_rhalf, datatype, op, schedule, true);
-        if (OPAL_UNLIKELY(OMPI_SUCCESS != res)) { goto cleanup_and_return; }
-        /* Send the right half to the left neighbor */
-        res = NBC_Sched_send((char *)rbuf + (ptrdiff_t)count_lhalf * extent,
-                             false, count_rhalf, datatype, rank - 1, schedule, true);
-        if (OPAL_UNLIKELY(OMPI_SUCCESS != res)) { goto cleanup_and_return; }
-        /* This process does not participate in recursive doubling phase */
-        vrank = -1;
-      } else {
-        /*
-         * Even process -- exchange with rank + 1
-         * Send the right half of the input vector to the right neighbor,
-         * Recv the left half of the input vector from the right neighbor
-         */
-        res = NBC_Sched_send((char *)rbuf + (ptrdiff_t)count_lhalf * extent,
-                             false, count_rhalf, datatype, rank + 1, schedule, false);
-        if (OPAL_UNLIKELY(OMPI_SUCCESS != res)) { goto cleanup_and_return; }
-        res = NBC_Sched_recv(tmp_buf, false, count_lhalf, datatype, rank + 1,
-                             schedule, true);
-        if (OPAL_UNLIKELY(OMPI_SUCCESS != res)) { goto cleanup_and_return; }
-        res = NBC_Sched_op(tmp_buf, false, rbuf, false, count_lhalf,
-                           datatype, op, schedule, true);
-        if (OPAL_UNLIKELY(OMPI_SUCCESS != res)) { goto cleanup_and_return; }
-        /* Recv the right half from the right neighbor */
-        res = NBC_Sched_recv((char *)rbuf + (ptrdiff_t)count_lhalf * extent,
-                             false, count_rhalf, datatype, rank + 1, schedule, true);
-        if (OPAL_UNLIKELY(OMPI_SUCCESS != res)) { goto cleanup_and_return; }
-        vrank = rank / 2;
-      }
-    } else { /* rank >= 2 * nprocs_rem */
-      vrank = rank - nprocs_rem;
-    }
-    /*
-     * Step 2. Reduce-scatter implemented with recursive vector halving and
-     * recursive distance doubling. We have p' = 2^{\floor{\log_2 p}}
-     * power-of-two number of processes with new ranks (vrank) and result in rbuf.
-     *
-     * The even-ranked processes send the right half of their buffer to rank + 1
-     * and the odd-ranked processes send the left half of their buffer to
-     * rank - 1. All processes then compute the reduction between the local
-     * buffer and the received buffer. In the next \log_2(p') - 1 steps, the
-     * buffers are recursively halved, and the distance is doubled. At the end,
-     * each of the p' processes has 1 / p' of the total reduction result.
-     */
-    rindex = malloc(sizeof(*rindex) * nsteps);
-    sindex = malloc(sizeof(*sindex) * nsteps);
-    rcount = malloc(sizeof(*rcount) * nsteps);
-    scount = malloc(sizeof(*scount) * nsteps);
-    if (NULL == rindex || NULL == sindex || NULL == rcount || NULL == scount) {
-      res = OMPI_ERR_OUT_OF_RESOURCE;
-      goto cleanup_and_return;
-    }
-    if (vrank != -1) {
-      step = 0;
-      wsize = count;
-      sindex[0] = rindex[0] = 0;
-      for (int mask = 1; mask < nprocs_pof2; mask <<= 1) {
-        /*
-         * On each iteration: rindex[step] = sindex[step] -- begining of the
-         * current window. Length of the current window is storded in wsize.
-         */
-        int vdest = vrank ^ mask;
-        /* Translate vdest virtual rank to real rank */
-        int dest = (vdest < nprocs_rem) ? vdest * 2 : vdest + nprocs_rem;
-        if (rank < dest) {
-          /*
-           * Recv into the left half of the current window, send the right
-           * half of the window to the peer (perform reduce on the left
-           * half of the current window)
-           */
-          rcount[step] = wsize / 2;
-          scount[step] = wsize - rcount[step];
-          sindex[step] = rindex[step] + rcount[step];
-        } else {
-          /*
-           * Recv into the right half of the current window, send the left
-           * half of the window to the peer (perform reduce on the right
-           * half of the current window)
-           */
-          scount[step] = wsize / 2;
-          rcount[step] = wsize - scount[step];
-          rindex[step] = sindex[step] + scount[step];
-        }
-        /* Send part of data from the rbuf, recv into the tmp_buf */
-        res = NBC_Sched_send((char *)rbuf + (ptrdiff_t)sindex[step] * extent,
-                             false, scount[step], datatype, dest, schedule, false);
-        if (OPAL_UNLIKELY(OMPI_SUCCESS != res)) { goto cleanup_and_return; }
-        res = NBC_Sched_recv((char *)tmp_buf + (ptrdiff_t)rindex[step] * extent,
-                             false, rcount[step], datatype, dest, schedule, true);
-        if (OPAL_UNLIKELY(OMPI_SUCCESS != res)) { goto cleanup_and_return; }
-        /* Local reduce: rbuf[] = tmp_buf[] <op> rbuf[] */
-        res = NBC_Sched_op((char *)tmp_buf + (ptrdiff_t)rindex[step] * extent,
-                           false, (char *)rbuf + (ptrdiff_t)rindex[step] * extent,
-                           false, rcount[step], datatype, op, schedule, true);
-        if (OPAL_UNLIKELY(OMPI_SUCCESS != res)) { goto cleanup_and_return; }
-        /* Move the current window to the received message */
-        if (step + 1 < nsteps) {
-          rindex[step + 1] = rindex[step];
-          sindex[step + 1] = rindex[step];
-          wsize = rcount[step];
-          step++;
-        }
-      }
-      /*
-       * Assertion: each process has 1 / p' of the total reduction result:
-       * rcount[nsteps - 1] elements in the rbuf[rindex[nsteps - 1], ...].
-       */
-      /*
-       * Step 3. Allgather by the recursive doubling algorithm.
-       * Each process has 1 / p' of the total reduction result:
-       * rcount[nsteps - 1] elements in the rbuf[rindex[nsteps - 1], ...].
-       * All exchanges are executed in reverse order relative
-       * to recursive doubling (previous step).
-       */
-      step = nsteps - 1;
-      for (int mask = nprocs_pof2 >> 1; mask > 0; mask >>= 1) {
-        int vdest = vrank ^ mask;
-        /* Translate vdest virtual rank to real rank */
-        int dest = (vdest < nprocs_rem) ? vdest * 2 : vdest + nprocs_rem;
-        /*
-         * Send rcount[step] elements from rbuf[rindex[step]...]
-         * Recv scount[step] elements to rbuf[sindex[step]...]
-         */
-        res = NBC_Sched_send((char *)rbuf + (ptrdiff_t)rindex[step] * extent,
-                             false, rcount[step], datatype, dest, schedule, false);
-        if (OPAL_UNLIKELY(OMPI_SUCCESS != res)) { goto cleanup_and_return; }
-        res = NBC_Sched_recv((char *)rbuf + (ptrdiff_t)sindex[step] * extent,
-                             false, scount[step], datatype, dest, schedule, true);
-        if (OPAL_UNLIKELY(OMPI_SUCCESS != res)) { goto cleanup_and_return; }
-        step--;
-      }
-    }
-    /*
-     * Step 4. Send total result to excluded odd ranks.
-     */
-    if (rank < 2 * nprocs_rem) {
-      if (rank % 2 != 0) {
-        /* Odd process -- recv result from rank - 1 */
-        res = NBC_Sched_recv(rbuf, false, count, datatype, rank - 1, schedule, false);
-        if (OPAL_UNLIKELY(OMPI_SUCCESS != res)) { goto cleanup_and_return; }
-      } else {
-        /* Even process -- send result to rank + 1 */
-        res = NBC_Sched_send(rbuf, false, count, datatype, rank + 1, schedule, false);
-        if (OPAL_UNLIKELY(OMPI_SUCCESS != res)) { goto cleanup_and_return; }
-      }
-    }
- cleanup_and_return:
-    if (NULL != rindex)
-      free(rindex);
-    if (NULL != sindex)
-      free(sindex);
-    if (NULL != rcount)
-      free(rcount);
-    if (NULL != scount)
-      free(scount);
-    return res;
-  }
-
-  /* int ompi_coll_libpnbc_osc_allreduce_init(const void* sendbuf, void* recvbuf, int count, */
-  /*                                     MPI_Datatype datatype, MPI_Op op, */
-  /*                                     struct ompi_communicator_t *comm, MPI_Info info, */
-  /*                                     ompi_request_t ** request, */
-  /*                                     struct mca_coll_base_module_2_3_0_t *module) { */
-  /*   int res = nbc_rallreduce_init(sendbuf, recvbuf, count, datatype, op, comm, request, module, true); */
-  /*   if (OPAL_UNLIKELY(OMPI_SUCCESS != res)) { */
-  /*     return res; */
-  /*   } */
-
-  /*   return OMPI_SUCCESS; */
-  /* } */
-
-  /* int ompi_coll_libpnbc_osc_allreduce_inter_init(const void* sendbuf, void* recvbuf, int count, */
-  /*                                           MPI_Datatype datatype, MPI_Op op, */
-  /*                                           struct ompi_communicator_t *comm, MPI_Info info, */
-  /*                                           ompi_request_t ** request, */
-  /*                                           struct mca_coll_base_module_2_3_0_t *module) { */
-  /*   int res = nbc_allreduce_inter_init(sendbuf, recvbuf, count, datatype, op, comm, request, */
-  /*                                      module, true); */
-  /*   if (OPAL_UNLIKELY(OMPI_SUCCESS != res)) { */
-  /*     return res; */
-  /*   } */
-  /* return OMPI_SUCCESS; */
-  /* } */
